requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0

psycopg2==2.9.11
psycopg-binary==3.3.2
pgvector==0.3.6
SQLAlchemy==2.0.45

zhipuai==2.1.5.20250825

langchain==1.2.0
langchain-community==0.4.1
langchain-core==1.2.5
langchain-postgres==0.0.16
langchain-openai==1.1.6
openai==2.14.0
tiktoken==0.12.0

python-dotenv==1.2.1
httpx==0.28.1
tenacity==9.1.2

sniffio==1.3.1
sqlmodel==0.0.27
PyYAML~=6.0.3

rapidocr_onnxruntime==1.4.4
opencv-python-headless==4.12.0.88
Pillow==11.1.0
python-multipart==0.0.21
starlette==0.50.0
opencv-python==4.11.0.86
numpy==2.2.6
matplotlib==3.10.8
requests==2.32.5